cropped_images/requirements.txt
pyheif 
Pillow